src/main/java/com/outgrowthsolutions/ogsrecipeapp/converters/CategoryToCategoryCommand.java
package com.outgrowthsolutions.ogsrecipeapp.converters;

import com.mongodb.lang.Nullable;
import com.outgrowthsolutions.ogsrecipeapp.commands.CategoryCommand;
import com.outgrowthsolutions.ogsrecipeapp.domain.Category;
import lombok.Synchronized;
import org.springframework.core.convert.converter.Converter;
import org.springframework.stereotype.Component;

@Component
public class CategoryToCategoryCommand implements Converter<Category, CategoryCommand> {
    @Synchronized
    @Override
    @Nullable
    public CategoryCommand convert(Category category) {
        if (category == null)
            return null;
        final CategoryCommand categoryCommand = new CategoryCommand();
        categoryCommand.setId(category.getId());
        categoryCommand.setDescription(category.getDescription());
        return categoryCommand;
    }
}
