src/test/java/com/outgrowthsolutions/ogsrecipeapp/converters/CategoryToCategoryCommandTest.java
package com.outgrowthsolutions.ogsrecipeapp.converters;

import com.outgrowthsolutions.ogsrecipeapp.commands.CategoryCommand;
import com.outgrowthsolutions.ogsrecipeapp.domain.Category;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;

class CategoryToCategoryCommandTest {

    CategoryToCategoryCommand converter;

    @BeforeEach
    void setUp() {
        converter = new CategoryToCategoryCommand();
    }

    @Test
    void convertNull() {
        assertNull(converter.convert(null));
    }

    @Test
    void convertEmpty() {
        assertNotNull(converter.convert(new Category()));
    }

    @Test
    void convert() {
        Category category = new Category();
        String id = "1L";
        category.setId(id);
        String description = "Indian";
        category.setDescription(description);

        CategoryCommand categoryCommand = converter.convert(category);
        assertNotNull(categoryCommand);
        assertEquals(id,categoryCommand.getId());
        assertEquals(description,categoryCommand.getDescription());
    }
}